+++ src/modules/chanmodes/delayjoin.c
@@ -31,7 +31,7 @@ int moded_quit(Client *client, MessageTag *mtags, const char *comment);
 int delayjoin_is_ok(Client *client, Channel *channel, char mode, const char *para, int checkt, int what);
 int moded_chanmode(Client *client, Channel *channel,
                    MessageTag *mtags, const char *modebuf, const char *parabuf, time_t sendts, int samode);
-int moded_prechanmsg(Client *client, Channel *channel, MessageTag *mtags, const char *text, SendType sendtype);
+int moded_prechanmsg(Client *client, Channel *channel, MessageTag **mtags, const char *text, SendType sendtype);
 const char *moded_serialize(ModData *m);
 void moded_unserialize(const char *str, ModData *m);
 
@@ -386,10 +386,10 @@ int moded_chanmode(Client *client, Channel *channel, MessageTag *recv_mtags, con
 	return 0;
 }
 
-int moded_prechanmsg(Client *client, Channel *channel, MessageTag *mtags, const char *text, SendType sendtype)
+int moded_prechanmsg(Client *client, Channel *channel, MessageTag **mtags, const char *text, SendType sendtype)
 {
 	if ((channel_is_delayed(channel) || channel_is_post_delayed(channel)) && (moded_user_invisible(client, channel)))
-		clear_user_invisible_announce(channel, client, mtags);
+		clear_user_invisible_announce(channel, client, *mtags);
 
 	return 0;
 }